adc+filter. not fast enough. Opti: known coeffs

+++ communication.X/filter.h
@@ -8,7 +8,7 @@
 #ifndef FILTER_H
 #define	FILTER_H
 void initFilter();
-void filterNewSample(float sample, float returnArray[2]);
+void filterNewSample(unsigned int sample, float returnArray[2]);
 
 //TEST
 void generateSineInput(int length, float frequency, int returnArray[]);
